+++ navigator/r2-navigator-swift/EPUB/EPUBSpreadView.swift
@@ -33,7 +33,7 @@ protocol EPUBSpreadViewDelegate: class {
     
 }
 
-class EPUBSpreadView: UIView, Loggable {
+class EPUBSpreadView: UIView, Loggable, PageView {
 
     weak var delegate: EPUBSpreadViewDelegate?
     let publication: Publication
@@ -277,11 +277,9 @@ class EPUBSpreadView: UIView, Loggable {
         // To be overridden in subclasses if the resource supports a progression.
         return 0
     }
-    
+
     func go(to location: PageLocation, completion: (() -> Void)?) {
-        // For fixed layout, there's only one page so location is not used. But this is overriden
-        // for reflowable resources.
-        completion?()
+        fatalError("go(to:completion:) must be implemented in subclasses")
     }
     
     enum Direction: CustomStringConvertible {
@@ -297,7 +295,7 @@ class EPUBSpreadView: UIView, Loggable {
     }
     
     func go(to direction: Direction, animated: Bool = false, completion: @escaping () -> Void = {}) -> Bool {
-        // The default implementation of a spread view consider that its content is entirely visible on screen.
+        // The default implementation of a spread view considers that its content is entirely visible on screen.
         return false
     }
 
@@ -392,23 +390,6 @@ class EPUBSpreadView: UIView, Loggable {
 
 }
 
-extension EPUBSpreadView: PageView {
-    
-    var positionCount: Int {
-        // Sum of the number of positions in all the resources of the spread.
-        return spread.links
-            .map {
-                if let index = publication.readingOrder.firstIndex(withHREF: $0.href) {
-                    return publication.positionsByReadingOrder[index].count
-                } else {
-                    return 0
-                }
-            }
-            .reduce(0, +)
-    }
-
-}
-
 // MARK: - WKScriptMessageHandler for handling incoming message from the javascript layer.
 extension EPUBSpreadView: WKScriptMessageHandler {
 
@@ -425,7 +406,7 @@ extension EPUBSpreadView: WKScriptMessageHandler {
 extension EPUBSpreadView: WKNavigationDelegate {
 
     func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
-        // Do not remove: overriden in subclasses.
+        // Do not remove: overridden in subclasses.
     }
 
     func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
@@ -460,7 +441,7 @@ extension EPUBSpreadView: UIScrollViewDelegate {
     }
 
     func scrollViewDidScroll(_ scrollView: UIScrollView) {
-        // Do not remove, overriden in subclasses.
+        // Do not remove, overridden in subclasses.
     }
 
 }