+++ navigator/r2-navigator-swift/Toolkit/CompletionList.swift
@@ -1,12 +1,7 @@
 //
-//  CompletionList.swift
-//  r2-navigator-swift
-//
-//  Created by Mickaël Menu on 21/02/2020.
-//
 //  Copyright 2020 Readium Foundation. All rights reserved.
-//  Use of this source code is governed by a BSD-style license which is detailed
-//  in the LICENSE file present in the project repository where this source code is maintained.
+//  Use of this source code is governed by the BSD-style license
+//  available in the top-level LICENSE file of the project.
 //
 
 import Foundation
@@ -30,17 +25,25 @@ final class CompletionList {
     /// Adds the given `completion` block the list.
     ///
     /// - Returns: A new block that will call all the registered completion blocks.
+    @discardableResult
     func add(_ completion: (() -> Void)?) -> () -> Void {
         if let completion = completion {
             blocks.append(completion)
         }
         
         return {
+            self.complete()
+        }
+    }
+
+    /// Calls all the registered completion blocks.
+    func complete() {
+        DispatchQueue.main.async {
             for block in self.blocks {
                 block()
             }
             self.blocks.removeAll()
         }
     }
-    
+
 }